Revert "fix: Align the core measure logic more with Windows"
