+++ src/SamplesApp/SamplesApp.UITests/Windows_UI_Xaml_Input/Nested_Sequence_Tests.cs
@@ -33,9 +33,6 @@ namespace SamplesApp.UITests.Windows_UI_Xaml_Input
 		[AutoRetry]
 		[ActivePlatforms(Platform.Android, Platform.iOS)]
 		[InjectedPointer(PointerDeviceType.Touch)]
-#if IS_RUNTIME_UI_TESTS
-		[Uno.UI.RuntimeTests.RequiresFullWindow]
-#endif
 		public async Task When_PressOnNestedAndReleaseOnContainer_Touch()
 		{
 			await RunAsync(_sample);
@@ -67,9 +64,6 @@ namespace SamplesApp.UITests.Windows_UI_Xaml_Input
 #if !__SKIA__
 		[Ignore("Inputs simulated by selenium are directly appreaing at the start location and wrongly inserting an exit.")]
 		//[ActivePlatforms(Platform.Browser)]
-#endif
-#if IS_RUNTIME_UI_TESTS
-		[Uno.UI.RuntimeTests.RequiresFullWindow]
 #endif
 		[InjectedPointer(PointerDeviceType.Mouse)]
 		public async Task When_PressOnNestedAndReleaseOnContainer_Mouse()
