+++ src/Uno.UI.RuntimeTests/Tests/Windows_UI_Xaml/Given_Control.cs
@@ -11,72 +11,12 @@ using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Shapes;
 using Windows.UI;
-using Windows.Foundation;
 
 namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml
 {
 	[TestClass]
-	public partial class Given_Control
+	public class Given_Control
 	{
-		private partial class CustomControl : Control
-		{
-			public Size AvailableSizePassedToMeasureOverride { get; private set; }
-			protected override Size MeasureOverride(Size availableSize)
-			{
-				AvailableSizePassedToMeasureOverride = availableSize;
-				return new(2000, 2000);
-			}
-		}
-
-		[TestMethod]
-		[RunsOnUIThread]
-		public async Task When_Limited_By_Available_Size_Before_Margin_Application()
-		{
-			// This assert makes sure that:
-			// 1. The availableSize passed to MeasureOverride is correct
-			// 2. DesiredSize is properly limited by the available size used for measure, before applying the margin.
-
-			// Note: This test might seem like it can be a unit test rather than a runtime test. But it's
-			// intentional to make it a runtime test. This is to make sure we test measure logic specific to each platform.
-
-			var SUT = new CustomControl
-			{
-				Margin = new Thickness(-70),
-			};
-
-			// availableSize passed to the core measure is 200x200
-			SUT.Measure(new Size(200, 200));
-
-			// The MeasureOverride is passed the size after applying margin, which is 340x340.
-			Assert.AreEqual(new Size(340, 340), SUT.AvailableSizePassedToMeasureOverride);
-
-			// The desiredSize is limited by the available size before applying the margin.
-			Assert.AreEqual(new Size(200, 200), SUT.DesiredSize);
-		}
-
-		[TestMethod]
-		[RunsOnUIThread]
-		public async Task When_Limited_By_Available_Size_After_Margin_Application()
-		{
-			// This assert makes sure that:
-			// 1. The availableSize passed to MeasureOverride is correct
-			// 2. DesiredSize is not limited by the available size used for measure, after applying the margin.
-
-			var SUT = new CustomControl
-			{
-				Margin = new Thickness(70),
-			};
-
-			// availableSize passed to the core measure is 200x200
-			SUT.Measure(new Size(200, 200));
-
-			// The MeasureOverride is passed the size after applying margin, which is 60x60.
-			Assert.AreEqual(new Size(60, 60), SUT.AvailableSizePassedToMeasureOverride);
-
-			// The desiredSize is NOT limited by the available size after applying the margin.
-			Assert.AreEqual(new Size(200, 200), SUT.DesiredSize);
-		}
-
 		[TestMethod]
 		[RunsOnUIThread]
 		public async Task When_SetChildTemplateUsingVisualState()
