+++ src/Uno.UI.RuntimeTests/Tests/Windows_UI_Xaml/Given_FrameworkElement.cs
@@ -36,7 +36,7 @@ using _View = Windows.UI.Xaml.UIElement;
 namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 {
 	[TestClass]
-	public partial class Given_FrameworkElement
+	public class Given_FrameworkElement
 	{
 #if __WASM__
 		// TODO Android does not handle measure invalidation properly
@@ -331,52 +331,31 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			});
 #endif
 
-		public partial class MyGrid : Grid
-		{
-			public Size AvailableSizeUsedForMeasure { get; private set; }
-			protected override Size MeasureOverride(Size availableSize)
-			{
-				AvailableSizeUsedForMeasure = availableSize;
-				return base.MeasureOverride(availableSize);
-			}
-		}
-
+#if __SKIA__
+		[Ignore("https://github.com/unoplatform/uno/issues/7271")]
+#endif
 		[TestMethod]
 		[RunsOnUIThread]
 #if __MACOS__
 		[Ignore("Currently fails on macOS, part of #9282 epic")]
 #endif
-#if __ANDROID__ || __IOS__
-		[Ignore("Layouter doesn't work properly")]
-#endif
 		public async Task When_MinWidth_SmallerThan_AvailableSize()
 		{
 			Border content = null;
 			ContentControl contentCtl = null;
-			MyGrid grid = null;
+			Grid grid = null;
 
 			content = new Border { Width = 100, Height = 15 };
 
 			contentCtl = new ContentControl { MinWidth = 110, Content = content };
 
-			grid = new MyGrid() { MinWidth = 120 };
+			grid = new Grid() { MinWidth = 120 };
 
 			grid.Children.Add(contentCtl);
 
 			grid.Measure(new Size(50, 50));
 
-			// This is matching Windows and is important to keep the behavior of this test like this.
-			Assert.AreEqual(new Size(120, 50), grid.AvailableSizeUsedForMeasure);
-
-#if __SKIA__
-			// https://github.com/unoplatform/uno/issues/7271
-			// This assert is NOT correct. The correct size is 50, 15
-			// It's happening because ContentControl is measuring incorrectly because GetFirstChild is returning null instead of the Border.
-			Assert.AreEqual(new Size(50, 0), grid.DesiredSize);
-#else
 			Assert.AreEqual(new Size(50, 15), grid.DesiredSize);
-#endif
-
 #if NETFX_CORE // Failing on WASM - https://github.com/unoplatform/uno/issues/2314
 			Assert.AreEqual(new Size(110, 15), contentCtl.DesiredSize);
 			Assert.AreEqual(new Size(100, 15), content.DesiredSize);
