+++ src/Uno.UI.RuntimeTests/Tests/Windows_UI_Xaml_Controls/Given_StackPanel.cs
@@ -150,24 +150,12 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 
 				if (!double.IsInfinity(maxConstraints.Width))
 				{
-#if WINDOWS_UWP || __CROSSRUNTIME__
-					Assert.AreEqual(constrained.ActualWidth, orientation == Orientation.Horizontal ? 1000 : maxConstraints.Width);
-#else
-					// TODO: Align Uno with Windows behavior.
 					Assert.AreEqual(constrained.ActualWidth, maxConstraints.Width);
-#endif
-
 					Assert.AreEqual(constrained.DesiredSize.Width, maxConstraints.Width);
 				}
 				if (!double.IsInfinity(maxConstraints.Height))
 				{
-#if WINDOWS_UWP || __CROSSRUNTIME__
-					Assert.AreEqual(constrained.ActualHeight, orientation == Orientation.Vertical ? 1000 : maxConstraints.Height);
-#else
-					// TODO: Align Uno with Windows behavior.
 					Assert.AreEqual(constrained.ActualHeight, maxConstraints.Height);
-#endif
-
 					Assert.AreEqual(constrained.DesiredSize.Height, maxConstraints.Height);
 				}
 			}
