+++ src/Uno.UI/UI/Xaml/Controls/Layouter/Layouter.cs
@@ -107,14 +107,6 @@ namespace Windows.UI.Xaml.Controls
 					.AtLeastZero()
 					.AtMost(maxSize);
 
-				// TODO: This commented code was done as part of aligning layouting on mobile platforms.
-				// We are reverting those changes as they require more changes, but keeping them
-				// commented for future reference.
-				//if (Panel is not ILayoutOptOut { ShouldUseMinSize: false })
-				//{
-				//	frameworkAvailableSize = frameworkAvailableSize.AtLeast(minSize);
-				//}
-
 				var desiredSize = MeasureOverride(frameworkAvailableSize);
 				LayoutInformation.SetAvailableSize(Panel, availableSize);
 
@@ -137,17 +129,8 @@ namespace Windows.UI.Xaml.Controls
 				_unclippedDesiredSize = desiredSize;
 
 				var clippedDesiredSize = desiredSize
-					// TODO: This commented code was done as part of aligning layouting on mobile platforms.
-					// We are reverting those changes as they require more changes, but keeping them
-					// commented for future reference.
-					//.AtMost(maxSize)
-					.AtMost(frameworkAvailableSize) // TODO: This line shouldn't be there (ie, frameworkAvailableSize should be maxSize).
+					.AtMost(frameworkAvailableSize)
 					.Add(marginSize)
-					// Making sure after adding margins that clipped DesiredSize is not bigger than the AvailableSize
-					// TODO: This commented code was done as part of aligning layouting on mobile platforms.
-					// We are reverting those changes as they require more changes, but keeping them
-					// commented for future reference.
-					//.AtMost(availableSize)
 					// Margin may be negative
 					.AtLeastZero();
 
@@ -226,35 +209,26 @@ namespace Windows.UI.Xaml.Controls
 
 				_logDebug?.Debug($"{this}: InnerArrangeCore({finalRect}) - allowClip={allowClipToSlot}, clippedArrangeSize={clippedArrangeSize}, _unclippedDesiredSize={_unclippedDesiredSize}, forcedClipping={needsClipToSlot}");
 
-				var arrangeSize = finalRect
-					.Size
-					.AtLeastZero(); // 0.0,0.0
-
 				if (allowClipToSlot && !needsClipToSlot)
 				{
 					if (IsLessThanAndNotCloseTo(clippedArrangeSize.Width, _unclippedDesiredSize.Width))
 					{
 						_logDebug?.Debug($"{this}: (arrangeSize.Width) {clippedArrangeSize.Width} < {_unclippedDesiredSize.Width}: NEEDS CLIPPING.");
 						needsClipToSlot = true;
-						// TODO: This commented code was done as part of aligning layouting on mobile platforms.
-						// We are reverting those changes as they require more changes, but keeping them
-						// commented for future reference.
-						//arrangeSize.Width = _unclippedDesiredSize.Width;
 					}
 
-					if (IsLessThanAndNotCloseTo(clippedArrangeSize.Height, _unclippedDesiredSize.Height))
+					else if (IsLessThanAndNotCloseTo(clippedArrangeSize.Height, _unclippedDesiredSize.Height))
 					{
 						_logDebug?.Debug($"{this}: (arrangeSize.Height) {clippedArrangeSize.Height} < {_unclippedDesiredSize.Height}: NEEDS CLIPPING.");
 						needsClipToSlot = true;
-						// TODO: This commented code was done as part of aligning layouting on mobile platforms.
-						// We are reverting those changes as they require more changes, but keeping them
-						// commented for future reference.
-						//arrangeSize.Height = _unclippedDesiredSize.Height;
 					}
 				}
 
-				var (_, maxSize) = this.Panel.GetMinMax();
-				//var marginSize = this.Panel.GetMarginSize();
+				var (minSize, maxSize) = this.Panel.GetMinMax();
+
+				var arrangeSize = finalRect
+					.Size
+					.AtLeastZero(); // 0.0,0.0
 
 				// We have to choose max between _unclippedDesiredSize and maxSize here, because
 				// otherwise setting of max property could cause arrange at less then _unclippedDesiredSize.
@@ -278,51 +252,11 @@ namespace Windows.UI.Xaml.Controls
 					}
 				}
 
-				var innerInkSize = ArrangeOverride(arrangeSize);
-				var clippedInkSize = innerInkSize.AtMost(maxSize);
-
-				// TODO: This commented code was done as part of aligning layouting on mobile platforms.
-				// We are reverting those changes as they require more changes, but keeping them
-				// commented for future reference.
-				//if (IsLessThanAndNotCloseTo(clippedInkSize.Width, innerInkSize.Width) || IsLessThanAndNotCloseTo(clippedInkSize.Height, innerInkSize.Height))
-				//{
-				//	needsClipToSlot = true;
-				//}
-
-				//var clientSize = finalRect.Size
-				//	.Subtract(marginSize)
-				//	.AtLeastZero();
-
-				//var (offset, overflow) = Panel.GetAlignmentOffset(clientSize, clippedInkSize);
-				//var margin = Panel.Margin;
-
-				//offset = new Point(
-				//	offset.X + finalRect.X + margin.Left,
-				//	offset.Y + finalRect.Y + margin.Top
-				//);
-
-				//if (overflow)
-				//{
-				//	needsClipToSlot = true;
-				//}
-
+				var renderSize = ArrangeOverride(arrangeSize);
 
 				if (_elementAsUIElement != null)
 				{
-					//_elementAsUIElement.LayoutSlotWithMarginsAndAlignments = new Rect(offset, innerInkSize);
-					//var layoutFrame = new Rect(offset, clippedInkSize);
-
-					// Calculate clipped frame.
-					//var clippedFrameWithParentOrigin = layoutFrame.IntersectWith(finalRect.DeflateBy(margin)) ?? Rect.Empty;
-
-					// Rebase the origin of the clipped frame to layout
-					//_elementAsUIElement.ClippedFrame = new Rect(
-					//	clippedFrameWithParentOrigin.X - layoutFrame.X,
-					//	clippedFrameWithParentOrigin.Y - layoutFrame.Y,
-					//	clippedFrameWithParentOrigin.Width,
-					//	clippedFrameWithParentOrigin.Height);
-
-					_elementAsUIElement.RenderSize = clippedInkSize; // TODO: This should be innerInkSize
+					_elementAsUIElement.RenderSize = renderSize.AtMost(maxSize);
 					_elementAsUIElement.NeedsClipToSlot = needsClipToSlot;
 					_elementAsUIElement.ApplyClip();
 
@@ -527,15 +461,12 @@ namespace Windows.UI.Xaml.Controls
 
 			LayoutInformation.SetLayoutSlot(view, frame);
 
-			// Note: This is not matching Windows.
-			// Applying alignments should depend on what ArrangeOverride returns (as in Skia and Wasm).
 			var (finalFrame, clippedFrame) = ApplyMarginAndAlignments(view, frame);
 			if (view is UIElement elt)
 			{
 				elt.LayoutSlotWithMarginsAndAlignments = finalFrame;
 				elt.ClippedFrame = clippedFrame;
 			}
-
 
 			ArrangeChildOverride(view, finalFrame);
 		}
@@ -553,6 +484,18 @@ namespace Windows.UI.Xaml.Controls
 
 		}
 #endif
+
+		protected Thickness MarginChild(View view)
+		{
+			if (view is IFrameworkElement frameworkElement)
+			{
+				return frameworkElement.Margin;
+			}
+			else
+			{
+				return Thickness.Empty;
+			}
+		}
 
 		protected abstract string Name { get; }
